requirements.txt
numpy
pandas
pdfplumber
streamlit
google.generativeai
pickle
textwrap